Concentrate debug/verbose functions in a single file (#310)
* buildbot trigger

* Searching for bugs...

* searching for bugs

* dont write packets

* Update README

* sync

* sync

* sync

* sync

* sync

* sync

* sync

* sync

* synv

* use central helper functions

* sync

* include programs_helper.c for windows

* windows buildfix

* nmake windows buildfix

* http client to be more verbose

* logic fix
