+++ programs/Makefile.am
@@ -78,45 +78,45 @@ noinst_PROGRAMS = \
     http_client \
     http_client_upcall
 
-test_libmgmt_SOURCES = test_libmgmt.c
+test_libmgmt_SOURCES = programs_helper.c test_libmgmt.c
 test_libmgmt_LDADD = ../usrsctplib/libusrsctp.la
-test_timer_SOURCES = test_timer.c
+test_timer_SOURCES = programs_helper.c test_timer.c
 test_timer_LDADD = ../usrsctplib/libusrsctp.la
-tsctp_SOURCES = tsctp.c
+tsctp_SOURCES = programs_helper.c tsctp.c
 tsctp_LDADD = ../usrsctplib/libusrsctp.la
-chargen_server_upcall_SOURCES = chargen_server_upcall.c
+chargen_server_upcall_SOURCES = programs_helper.c chargen_server_upcall.c
 chargen_server_upcall_LDADD = ../usrsctplib/libusrsctp.la
-daytime_server_SOURCES = daytime_server.c
+daytime_server_SOURCES = programs_helper.c daytime_server.c
 daytime_server_LDADD = ../usrsctplib/libusrsctp.la
-daytime_server_upcall_SOURCES = daytime_server_upcall.c
+daytime_server_upcall_SOURCES = programs_helper.c daytime_server_upcall.c
 daytime_server_upcall_LDADD = ../usrsctplib/libusrsctp.la
-discard_server_SOURCES = discard_server.c
+discard_server_SOURCES = programs_helper.c discard_server.c
 discard_server_LDADD = ../usrsctplib/libusrsctp.la
-discard_server_upcall_SOURCES = discard_server_upcall.c
+discard_server_upcall_SOURCES = programs_helper.c discard_server_upcall.c
 discard_server_upcall_LDADD = ../usrsctplib/libusrsctp.la
-echo_server_SOURCES = echo_server.c
+echo_server_SOURCES = programs_helper.c echo_server.c
 echo_server_LDADD = ../usrsctplib/libusrsctp.la
-echo_server_upcall_SOURCES = echo_server_upcall.c
+echo_server_upcall_SOURCES = programs_helper.c echo_server_upcall.c
 echo_server_upcall_LDADD = ../usrsctplib/libusrsctp.la
-client_SOURCES = client.c
+client_SOURCES = programs_helper.c client.c
 client_LDADD = ../usrsctplib/libusrsctp.la
-client_upcall_SOURCES = client_upcall.c
+client_upcall_SOURCES = programs_helper.c client_upcall.c
 client_upcall_LDADD = ../usrsctplib/libusrsctp.la
-rtcweb_SOURCES = rtcweb.c
+rtcweb_SOURCES = programs_helper.c rtcweb.c
 rtcweb_LDADD = ../usrsctplib/libusrsctp.la
-ekr_server_SOURCES = ekr_server.c
+ekr_server_SOURCES = programs_helper.c ekr_server.c
 ekr_server_LDADD = ../usrsctplib/libusrsctp.la
-ekr_client_SOURCES = ekr_client.c
+ekr_client_SOURCES = programs_helper.c ekr_client.c
 ekr_client_LDADD = ../usrsctplib/libusrsctp.la
-ekr_peer_SOURCES = ekr_peer.c
+ekr_peer_SOURCES = programs_helper.c ekr_peer.c
 ekr_peer_LDADD = ../usrsctplib/libusrsctp.la
-ekr_loop_SOURCES = ekr_loop.c
+ekr_loop_SOURCES = programs_helper.c ekr_loop.c
 ekr_loop_LDADD = ../usrsctplib/libusrsctp.la
-ekr_loop_offload_SOURCES = ekr_loop_offload.c
+ekr_loop_offload_SOURCES = programs_helper.c ekr_loop_offload.c
 ekr_loop_offload_LDADD = ../usrsctplib/libusrsctp.la
-ekr_loop_upcall_SOURCES = ekr_loop_upcall.c
+ekr_loop_upcall_SOURCES = programs_helper.c ekr_loop_upcall.c
 ekr_loop_upcall_LDADD = ../usrsctplib/libusrsctp.la
-http_client_SOURCES = http_client.c
+http_client_SOURCES = programs_helper.c http_client.c
 http_client_LDADD = ../usrsctplib/libusrsctp.la
-http_client_upcall_SOURCES = http_client_upcall.c
+http_client_upcall_SOURCES = programs_helper.c http_client_upcall.c
 http_client_upcall_LDADD = ../usrsctplib/libusrsctp.la
