+++ programs/Makefile.nmake
@@ -38,6 +38,7 @@ LINKFLAGS=/LIBPATH:..\usrsctplib usrsctp.lib
 	$(CC) $(CVARSDLL) $(CFLAGS) -Fd.\ -c $<
 
 all: \
+	programs_helper.obj \
 	client \
 	client_upcall \
 	daytime_server \
@@ -57,82 +58,85 @@ all: \
 	test_libmgmt \
 	http_client \
 	http_client_upcall
+	
+programs_helper.obj : programs_helper.c programs_helper.h
+	cl $(CVARSDLL) $(CFLAGS) -c programs_helper.c
 
 client:
 	$(CC) $(CFLAGS) $(CVARSDLL) -c client.c
-	link -out:client.exe client.obj $(LINKFLAGS)
+	link -out:client.exe client.obj programs_helper.obj $(LINKFLAGS)
 
 client_upcall:
 	$(CC) $(CFLAGS) $(CVARSDLL) -c client_upcall.c
-	link -out:client_upcall.exe client_upcall.obj $(LINKFLAGS)
+	link -out:client_upcall.exe client_upcall.obj programs_helper.obj $(LINKFLAGS)
 
 daytime_server:
 	$(CC) $(CFLAGS) $(CVARSDLL) -c daytime_server.c
-	link -out:daytime_server.exe daytime_server.obj $(LINKFLAGS)
+	link -out:daytime_server.exe daytime_server.obj programs_helper.obj $(LINKFLAGS)
 
 daytime_server_upcall:
 	$(CC) $(CFLAGS) $(CVARSDLL) -c daytime_server_upcall.c
-	link -out:daytime_server_upcall.exe daytime_server_upcall.obj $(LINKFLAGS)
+	link -out:daytime_server_upcall.exe daytime_server_upcall.obj programs_helper.obj $(LINKFLAGS)
 
 discard_server:
 	$(CC) $(CFLAGS) $(CVARSDLL) -c discard_server.c
-	link -out:discard_server.exe discard_server.obj $(LINKFLAGS)
+	link -out:discard_server.exe discard_server.obj programs_helper.obj $(LINKFLAGS)
 
 discard_server_upcall:
 	$(CC) $(CFLAGS) $(CVARSDLL) -c discard_server_upcall.c
-	link -out:discard_server_upcall.exe discard_server_upcall.obj $(LINKFLAGS)
+	link -out:discard_server_upcall.exe discard_server_upcall.obj programs_helper.obj $(LINKFLAGS)
 
 echo_server:
 	$(CC) $(CFLAGS) $(CVARSDLL) -c echo_server.c
-	link -out:echo_server.exe echo_server.obj $(LINKFLAGS)
+	link -out:echo_server.exe echo_server.obj programs_helper.obj $(LINKFLAGS)
 
 echo_server_upcall:
 	$(CC) $(CFLAGS) $(CVARSDLL) -c echo_server_upcall.c
-	link -out:echo_server_upcall.exe echo_server_upcall.obj $(LINKFLAGS)
+	link -out:echo_server_upcall.exe echo_server_upcall.obj programs_helper.obj $(LINKFLAGS)
 
 tsctp:
 	$(CC) $(CFLAGS) $(CVARSDLL) -c tsctp.c
-	link -out:tsctp.exe tsctp.obj $(LINKFLAGS)
+	link -out:tsctp.exe tsctp.obj programs_helper.obj $(LINKFLAGS)
 
 rtcweb:
 	$(CC) $(CFLAGS) $(CVARSDLL) -c rtcweb.c
-	link -out:rtcweb.exe rtcweb.obj $(LINKFLAGS)
+	link -out:rtcweb.exe rtcweb.obj programs_helper.obj $(LINKFLAGS)
 
 ekr_client:
 	$(CC) $(CFLAGS) $(CVARSDLL) -c ekr_client.c
-	link -out:ekr_client.exe ekr_client.obj $(LINKFLAGS)
+	link -out:ekr_client.exe ekr_client.obj programs_helper.obj $(LINKFLAGS)
 
 ekr_server:
 	$(CC) $(CFLAGS) $(CVARSDLL) -c ekr_server.c
-	link -out:ekr_server.exe ekr_server.obj $(LINKFLAGS)
+	link -out:ekr_server.exe ekr_server.obj programs_helper.obj $(LINKFLAGS)
 
 ekr_peer:
 	$(CC) $(CFLAGS) $(CVARSDLL) -c ekr_peer.c
-	link -out:ekr_peer.exe ekr_peer.obj $(LINKFLAGS)
+	link -out:ekr_peer.exe ekr_peer.obj programs_helper.obj $(LINKFLAGS)
 
 ekr_loop:
 	$(CC) $(CFLAGS) $(CVARSDLL) -c ekr_loop.c
-	link -out:ekr_loop.exe ekr_loop.obj $(LINKFLAGS)
+	link -out:ekr_loop.exe ekr_loop.obj programs_helper.obj $(LINKFLAGS)
 
 ekr_loop_offload:
 	$(CC) $(CFLAGS) $(CVARSDLL) -c ekr_loop_offload.c
-	link -out:ekr_loop_offload.exe ekr_loop_offload.obj $(LINKFLAGS)
+	link -out:ekr_loop_offload.exe ekr_loop_offload.obj programs_helper.obj $(LINKFLAGS)
 
 ekr_loop_upcall:
 	$(CC) $(CFLAGS) $(CVARSDLL) -c ekr_loop_upcall.c
-	link -out:ekr_loop_upcall.exe ekr_loop_upcall.obj $(LINKFLAGS)
+	link -out:ekr_loop_upcall.exe ekr_loop_upcall.obj programs_helper.obj $(LINKFLAGS)
 
 test_libmgmt:
 	$(CC) $(CFLAGS) $(CVARSDLL) -c test_libmgmt.c
-	link -out:test_libmgmt.exe test_libmgmt.obj $(LINKFLAGS)
+	link -out:test_libmgmt.exe test_libmgmt.obj programs_helper.obj $(LINKFLAGS)
 
 http_client:
 	$(CC) $(CFLAGS) $(CVARSDLL) -c http_client.c
-	link -out:http_client.exe http_client.obj $(LINKFLAGS)
+	link -out:http_client.exe http_client.obj programs_helper.obj $(LINKFLAGS)
 
 http_client_upcall:
 	$(CC) $(CFLAGS) $(CVARSDLL) -c http_client_upcall.c
-	link -out:http_client_upcall.exe http_client_upcall.obj $(LINKFLAGS)
+	link -out:http_client_upcall.exe http_client_upcall.obj programs_helper.obj $(LINKFLAGS)
 
 clean:
 	del /F client.exe
