+++ programs/chargen_server_upcall.c
@@ -49,6 +49,7 @@
 #include <arpa/inet.h>
 #endif
 #include <usrsctp.h>
+#include "programs_helper.h"
 
 #define BUFFERSIZE 10240
 #define PORT 19
@@ -146,16 +147,6 @@ handle_upcall(struct socket *upcall_socket, void *upcall_data, int upcall_flags)
 	}
 
 	return;
-}
-
-void
-debug_printf(const char *format, ...)
-{
-	va_list ap;
-
-	va_start(ap, format);
-	vprintf(format, ap);
-	va_end(ap);
 }
 
 int
