+++ programs/client.c
@@ -51,6 +51,7 @@
 #include <io.h>
 #endif
 #include <usrsctp.h>
+#include "programs_helper.h"
 
 int done = 0;
 
@@ -58,205 +59,7 @@ int done = 0;
 typedef char* caddr_t;
 #endif
 
-static void
-handle_association_change_event(struct sctp_assoc_change *sac)
-{
-	unsigned int i, n;
 
-	printf("Association change ");
-	switch (sac->sac_state) {
-	case SCTP_COMM_UP:
-		printf("SCTP_COMM_UP");
-		break;
-	case SCTP_COMM_LOST:
-		printf("SCTP_COMM_LOST");
-		break;
-	case SCTP_RESTART:
-		printf("SCTP_RESTART");
-		break;
-	case SCTP_SHUTDOWN_COMP:
-		printf("SCTP_SHUTDOWN_COMP");
-		break;
-	case SCTP_CANT_STR_ASSOC:
-		printf("SCTP_CANT_STR_ASSOC");
-		break;
-	default:
-		printf("UNKNOWN");
-		break;
-	}
-	printf(", streams (in/out) = (%u/%u)",
-	       sac->sac_inbound_streams, sac->sac_outbound_streams);
-	n = sac->sac_length - sizeof(struct sctp_assoc_change);
-	if (((sac->sac_state == SCTP_COMM_UP) ||
-	     (sac->sac_state == SCTP_RESTART)) && (n > 0)) {
-		printf(", supports");
-		for (i = 0; i < n; i++) {
-			switch (sac->sac_info[i]) {
-			case SCTP_ASSOC_SUPPORTS_PR:
-				printf(" PR");
-				break;
-			case SCTP_ASSOC_SUPPORTS_AUTH:
-				printf(" AUTH");
-				break;
-			case SCTP_ASSOC_SUPPORTS_ASCONF:
-				printf(" ASCONF");
-				break;
-			case SCTP_ASSOC_SUPPORTS_MULTIBUF:
-				printf(" MULTIBUF");
-				break;
-			case SCTP_ASSOC_SUPPORTS_RE_CONFIG:
-				printf(" RE-CONFIG");
-				break;
-			default:
-				printf(" UNKNOWN(0x%02x)", sac->sac_info[i]);
-				break;
-			}
-		}
-	} else if (((sac->sac_state == SCTP_COMM_LOST) ||
-	            (sac->sac_state == SCTP_CANT_STR_ASSOC)) && (n > 0)) {
-		printf(", ABORT =");
-		for (i = 0; i < n; i++) {
-			printf(" 0x%02x", sac->sac_info[i]);
-		}
-	}
-	printf(".\n");
-	if ((sac->sac_state == SCTP_CANT_STR_ASSOC) ||
-	    (sac->sac_state == SCTP_SHUTDOWN_COMP) ||
-	    (sac->sac_state == SCTP_COMM_LOST)) {
-		exit(0);
-	}
-	return;
-}
-
-static void
-handle_peer_address_change_event(struct sctp_paddr_change *spc)
-{
-	char addr_buf[INET6_ADDRSTRLEN];
-	const char *addr;
-	struct sockaddr_in *sin;
-	struct sockaddr_in6 *sin6;
-	struct sockaddr_conn *sconn;
-
-	switch (spc->spc_aaddr.ss_family) {
-	case AF_INET:
-		sin = (struct sockaddr_in *)&spc->spc_aaddr;
-		addr = inet_ntop(AF_INET, &sin->sin_addr, addr_buf, INET_ADDRSTRLEN);
-		break;
-	case AF_INET6:
-		sin6 = (struct sockaddr_in6 *)&spc->spc_aaddr;
-		addr = inet_ntop(AF_INET6, &sin6->sin6_addr, addr_buf, INET6_ADDRSTRLEN);
-		break;
-	case AF_CONN:
-		sconn = (struct sockaddr_conn *)&spc->spc_aaddr;
-#ifdef _WIN32
-		_snprintf(addr_buf, INET6_ADDRSTRLEN, "%p", sconn->sconn_addr);
-#else
-		snprintf(addr_buf, INET6_ADDRSTRLEN, "%p", sconn->sconn_addr);
-#endif
-		addr = addr_buf;
-		break;
-	default:
-#ifdef _WIN32
-		_snprintf(addr_buf, INET6_ADDRSTRLEN, "Unknown family %d", spc->spc_aaddr.ss_family);
-#else
-		snprintf(addr_buf, INET6_ADDRSTRLEN, "Unknown family %d", spc->spc_aaddr.ss_family);
-#endif
-		addr = addr_buf;
-		break;
-	}
-	printf("Peer address %s is now ", addr);
-	switch (spc->spc_state) {
-	case SCTP_ADDR_AVAILABLE:
-		printf("SCTP_ADDR_AVAILABLE");
-		break;
-	case SCTP_ADDR_UNREACHABLE:
-		printf("SCTP_ADDR_UNREACHABLE");
-		break;
-	case SCTP_ADDR_REMOVED:
-		printf("SCTP_ADDR_REMOVED");
-		break;
-	case SCTP_ADDR_ADDED:
-		printf("SCTP_ADDR_ADDED");
-		break;
-	case SCTP_ADDR_MADE_PRIM:
-		printf("SCTP_ADDR_MADE_PRIM");
-		break;
-	case SCTP_ADDR_CONFIRMED:
-		printf("SCTP_ADDR_CONFIRMED");
-		break;
-	default:
-		printf("UNKNOWN");
-		break;
-	}
-	printf(" (error = 0x%08x).\n", spc->spc_error);
-	return;
-}
-
-static void
-handle_send_failed_event(struct sctp_send_failed_event *ssfe)
-{
-	size_t i, n;
-
-	if (ssfe->ssfe_flags & SCTP_DATA_UNSENT) {
-		printf("Unsent ");
-	}
-	if (ssfe->ssfe_flags & SCTP_DATA_SENT) {
-		printf("Sent ");
-	}
-	if (ssfe->ssfe_flags & ~(SCTP_DATA_SENT | SCTP_DATA_UNSENT)) {
-		printf("(flags = %x) ", ssfe->ssfe_flags);
-	}
-	printf("message with PPID = %u, SID = %u, flags: 0x%04x due to error = 0x%08x",
-	       ntohl(ssfe->ssfe_info.snd_ppid), ssfe->ssfe_info.snd_sid,
-	       ssfe->ssfe_info.snd_flags, ssfe->ssfe_error);
-	n = ssfe->ssfe_length - sizeof(struct sctp_send_failed_event);
-	for (i = 0; i < n; i++) {
-		printf(" 0x%02x", ssfe->ssfe_data[i]);
-	}
-	printf(".\n");
-	return;
-}
-
-static void
-handle_notification(union sctp_notification *notif, size_t n)
-{
-	if (notif->sn_header.sn_length != (uint32_t)n) {
-		return;
-	}
-	switch (notif->sn_header.sn_type) {
-	case SCTP_ASSOC_CHANGE:
-		handle_association_change_event(&(notif->sn_assoc_change));
-		break;
-	case SCTP_PEER_ADDR_CHANGE:
-		handle_peer_address_change_event(&(notif->sn_paddr_change));
-		break;
-	case SCTP_REMOTE_ERROR:
-		break;
-	case SCTP_SHUTDOWN_EVENT:
-		break;
-	case SCTP_ADAPTATION_INDICATION:
-		break;
-	case SCTP_PARTIAL_DELIVERY_EVENT:
-		break;
-	case SCTP_AUTHENTICATION_EVENT:
-		break;
-	case SCTP_SENDER_DRY_EVENT:
-		break;
-	case SCTP_NOTIFICATIONS_STOPPED_EVENT:
-		break;
-	case SCTP_SEND_FAILED_EVENT:
-		handle_send_failed_event(&(notif->sn_send_failed_event));
-		break;
-	case SCTP_STREAM_RESET_EVENT:
-		break;
-	case SCTP_ASSOC_RESET_EVENT:
-		break;
-	case SCTP_STREAM_CHANGE_EVENT:
-		break;
-	default:
-		break;
-	}
-}
 
 static int
 receive_cb(struct socket *sock, union sctp_sockstore addr, void *data,
@@ -280,16 +83,6 @@ receive_cb(struct socket *sock, union sctp_sockstore addr, void *data,
 		free(data);
 	}
 	return (1);
-}
-
-void
-debug_printf(const char *format, ...)
-{
-	va_list ap;
-
-	va_start(ap, format);
-	vprintf(format, ap);
-	va_end(ap);
 }
 
 int
