+++ programs/daytime_server.c
@@ -52,20 +52,11 @@
 #include <arpa/inet.h>
 #endif
 #include <usrsctp.h>
+#include "programs_helper.h"
 
 #define PORT 13
 #define DAYTIME_PPID 40
 #define SLEEP 1
-
-void
-debug_printf(const char *format, ...)
-{
-	va_list ap;
-
-	va_start(ap, format);
-	vprintf(format, ap);
-	va_end(ap);
-}
 
 int
 main(int argc, char *argv[])
