+++ programs/daytime_server_upcall.c
@@ -48,6 +48,7 @@
 #include <arpa/inet.h>
 #endif
 #include <usrsctp.h>
+#include "programs_helper.h"
 
 #define DAYTIME_PPID 40
 #define PORT 13
@@ -80,16 +81,6 @@ handle_accept(struct socket *sock, void *data, int flags)
 		usrsctp_sendv(conn_sock, buffer, strlen(buffer), NULL, 0, (void *)&sndinfo,
 		              (socklen_t)sizeof(struct sctp_sndinfo), SCTP_SENDV_SNDINFO, 0);
 		usrsctp_close(conn_sock);
-}
-
-void
-debug_printf(const char *format, ...)
-{
-	va_list ap;
-
-	va_start(ap, format);
-	vprintf(format, ap);
-	va_end(ap);
 }
 
 int
