+++ programs/ekr_client.c
@@ -48,6 +48,7 @@
 #include <ws2tcpip.h>
 #endif
 #include <usrsctp.h>
+#include "programs_helper.h"
 
 #define MAX_PACKET_SIZE (1<<16)
 #define BUFFER_SIZE 80
@@ -150,16 +151,6 @@ receive_cb(struct socket *sock, union sctp_sockstore addr, void *data,
 		usrsctp_close(sock);
 	}
 	return (1);
-}
-
-void
-debug_printf(const char *format, ...)
-{
-	va_list ap;
-
-	va_start(ap, format);
-	vprintf(format, ap);
-	va_end(ap);
 }
 
 int
