+++ programs/ekr_server.c
@@ -49,6 +49,7 @@
 #include <ws2tcpip.h>
 #endif
 #include <usrsctp.h>
+#include "programs_helper.h"
 
 #define PORT 5001
 #define MAX_PACKET_SIZE (1<<16)
@@ -150,16 +151,6 @@ receive_cb(struct socket *s, union sctp_sockstore addr, void *data,
 		usrsctp_close(s);
 	}
 	return (1);
-}
-
-void
-debug_printf(const char *format, ...)
-{
-	va_list ap;
-
-	va_start(ap, format);
-	vprintf(format, ap);
-	va_end(ap);
 }
 
 int
