+++ programs/rtcweb.c
@@ -56,6 +56,7 @@
 #include <string.h>
 #include <errno.h>
 #include <usrsctp.h>
+#include "programs_helper.h"
 
 #define LINE_LENGTH (1024)
 #define BUFFER_SIZE (1<<16)
@@ -1130,7 +1131,7 @@ handle_send_failed_event(struct sctp_send_failed_event *ssfe)
 }
 
 static void
-handle_notification(struct peer_connection *pc, union sctp_notification *notif, size_t n)
+handle_notification_rtcweb(struct peer_connection *pc, union sctp_notification *notif, size_t n)
 {
 	if (notif->sn_header.sn_length != (uint32_t)n) {
 		return;
@@ -1292,23 +1293,13 @@ receive_cb(struct socket *sock, union sctp_sockstore addr, void *data,
 	if (data) {
 		lock_peer_connection(pc);
 		if (flags & MSG_NOTIFICATION) {
-			handle_notification(pc, (union sctp_notification *)data, datalen);
+			handle_notification_rtcweb(pc, (union sctp_notification *)data, datalen);
 		} else {
 			handle_message(pc, data, datalen, ntohl(rcv.rcv_ppid), rcv.rcv_sid);
 		}
 		unlock_peer_connection(pc);
 	}
 	return (1);
-}
-
-void
-debug_printf(const char *format, ...)
-{
-	va_list ap;
-
-	va_start(ap, format);
-	vprintf(format, ap);
-	va_end(ap);
 }
 
 int
