+++ programs/tsctp_upcall.c
@@ -56,6 +56,7 @@
 #include <getopt.h>
 #endif
 #include <usrsctp.h>
+#include "programs_helper.h"
 
 #define TSCTP_CLIENT 1
 #define TSCTP_SERVER 2
@@ -357,16 +358,6 @@ handle_upcall(struct socket *upcall_socket, void *upcall_data, int upcall_flags)
 	}
 
 	return;
-}
-
-void
-debug_printf(const char *format, ...)
-{
-	va_list ap;
-
-	va_start(ap, format);
-	vprintf(format, ap);
-	va_end(ap);
 }
 
 int main(int argc, char **argv)
